Refactored to use Decimal instead of double.

+++ lib/src/utils/utilities.dart
@@ -52,7 +52,7 @@ class _Utilities {
     this.settings = settings ?? MoneyFormatterSettings();
   }
 
-  double amount;
+  Decimal amount;
 
   MoneyFormatterSettings settings;
